auth tests moved to api folder
