+++ test/api/auth-user.test.js
@@ -2,7 +2,7 @@ const chai = require('chai');
 const chaiHttp = require('chai-http');
 chai.use(chaiHttp);
 const assert = chai.assert;
-const app = require('../lib/app');
+const app = require('../../lib/app');
 
 describe('authenticate credentials', () => {
